Use ExternalProject_add to download captone only
Attempting to use funchook in an Android gradle build results in ninja:
error: 'lib/libcapstone.a', needed by 'xyz.so', missing and no known
rule to make it. Switch out the clunky use of ExternalProject_add for
downloading and building capstone in favour of using it only for
downloading followed by add_subdirectory.